Sample Code from Report a Sea Turtle app
Includes java code from two activities of Report a Sea Turtle app as
well as php code (insert_sighting.php) that interacts with MySQL
database.

index.php is the code for a webpage that contains the .apk and shows
screenshots of the app as well as the latest 20 entries in the SQL
database.